refactor(RacingCar): 전진 시도 전략패턴 도입
자동차의 전진 시도에 대해, 시도 방법을 전략패턴을 도입
 - 다양한 전략을 추가할 수 있으나, 현재는 랜덤으로 4이상인지 확인하는 전략(RandomStratedy) 사용
 - 관련 코드 수정
