+++ src/main/java/racingcar/model/stratedy/MovableStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.model.stratedy;
+
+public interface MovableStrategy {
+    boolean isMove();
+}
